/transaction numbers now being parsed as doubles to retain decimals

+++ BackEnd/transmicroservices/src/main/java/com/rmit/sept/bk_transservices/transmicroservices/model/TransApiBody.java
@@ -6,5 +6,5 @@ import java.util.Map;
 public class TransApiBody
 {
     public Long buyer_id;
-    public List<Map<String, Long>> listings;
+    public List<Map<String, Double>> listings;
 }
